requirements.txt
numpy>=1.23.2
pandas>=1.5.0
redis>=4.3.4
pyyaml
coloredlogs
